Refactor ETL into package structure
