+++ backend/app/etl_runner/__init__.py
@@ -13,10 +13,10 @@ from .metadata import (
 )
 from .runner import (
     DataLoader,
-    _RunEtlFactory,
-    _RunEtlFunc,
     _load_run_etl,
     _run_etl_with_retries,
+    _RunEtlFactory,
+    _RunEtlFunc,
     _utc_now,
     start_etl_job,
 )
